fix(playerbot): Move LootDistribution to forward decl, add include guards, fix GuildIntegration types
- Fixed LootDistribution redefinition by using forward declaration in GameSystemsManager.h and moving include to .cpp
- Added explicit include guards to LootDistribution.h in addition to #pragma once
- Fixed GuildIntegration override errors by moving GuildChatMessage, GuildProfile, GuildParticipation, and related enums to IGuildIntegration.h interface
- Removed duplicate destructor in GuildIntegration.h
- Moved GuildActivityType, GuildChatStyle, and GuildRole enums to interface for proper type matching
- All structs used in interface method signatures now have full definitions in interface headers
