+++ src/modules/Playerbot/Core/Managers/GameSystemsManager.cpp
@@ -24,6 +24,9 @@
 #include "Professions/AuctionMaterialsBridge.h"
 // #include "Spatial/BlackboardManager.h"  // TODO: File does not exist
 
+// Manager implementations (for unique_ptr destruction)
+#include "Social/LootDistribution.h"
+
 #include <set>
 
 namespace Playerbot
